Fixed error with strings and const char* that crashes the Arduino programm

+++ ESP32/REST_API_JSON_Serial_Wifi_motor_PS3_v0/motor_parameters.h
@@ -12,9 +12,9 @@
 static const bool ENABLE_JOYSTICK = true;
 constexpr int joystickSensitivity = 75; // for comparison with number in the range of 0-512
 
-int POSITION_MOTOR_X = 0;
-int POSITION_MOTOR_Y = 0;
-int POSITION_MOTOR_Z = 0;
+long POSITION_MOTOR_X = 0;
+long POSITION_MOTOR_Y = 0;
+long POSITION_MOTOR_Z = 0;
 
 
 // Motorized stage
